Services/stt/requirements.txt
faster-whisper==1.2.1
ctranslate2==4.7.1
av==16.1.0
numpy==2.4.2
fastapi==0.128.7
uvicorn==0.40.0
python-multipart==0.0.22
prometheus-fastapi-instrumentator==7.1.0
httpx==0.28.1
pydantic-settings==2.13.1